Refactor MultiplayerManager: improve code structure and organization for better readability and maintainability

+++ SOS/include/network/MultiplayerManager.h
@@ -43,7 +43,7 @@ public:
     void sendPlayerAction(int actionType);
     
     // Send enemy state update to server (e.g., when enemy dies)
-    void sendEnemyStateUpdate(const std::string& enemyId, bool isDead, int currentHealth);
+    void sendEnemyStateUpdate(const std::string& enemyId, bool isDead, int16_t currentHealth);
     void handleEnemyStateMessage(const NetworkMessage& message);
 
     // Check if connected to server
